+++ src/app/(dashboard)/clients/page.tsx
@@ -35,7 +35,6 @@ export default async function ClientsPage({
   if (searchParams.search) {
     andConditions.push({
       OR: [
-        { companyName: { contains: searchParams.search, mode: "insensitive" } },
         { firstName: { contains: searchParams.search, mode: "insensitive" } },
         { lastName: { contains: searchParams.search, mode: "insensitive" } },
         { email: { contains: searchParams.search, mode: "insensitive" } },